store/requirements.txt
Babel==1.3
Django==1.8.3
django-appconf==1.0.1
django-compressor==1.5
django-extra-views==0.6.4
django-haystack==2.3.1
django-oscar==1.1.1
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.3
factory-boy==2.4.1
mock==1.0.1
phonenumbers==6.3.0
Pillow==2.7.0
psycopg2==2.6.1
purl==1.1
pytz==2015.4
six==1.9.0
sorl-thumbnail==12.2
Unidecode==0.4.18
